+++ src/Models/ZoneMemberProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ZoneMemberProxy class.
+ *
+ * @copyright   Copyright (c) 2023 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-02-26
+ *
+ */
+
+namespace Konekt\Address\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class ZoneMemberProxy extends ModelProxy
+{
+}