+++ src/Providers/ModuleServiceProvider.php
@@ -11,15 +11,23 @@
 
 namespace Konekt\Address\Providers;
 
+use Illuminate\Database\Eloquent\Relations\Relation;
 use Konekt\Address\Models\Address;
 use Konekt\Address\Models\AddressType;
 use Konekt\Address\Models\Country;
+use Konekt\Address\Models\CountryProxy;
 use Konekt\Address\Models\Gender;
 use Konekt\Address\Models\NameOrder;
 use Konekt\Address\Models\Organization;
 use Konekt\Address\Models\Person;
 use Konekt\Address\Models\Province;
+use Konekt\Address\Models\ProvinceProxy;
 use Konekt\Address\Models\ProvinceType;
+use Konekt\Address\Models\Zone;
+use Konekt\Address\Models\ZoneMember;
+use Konekt\Address\Models\ZoneMemberType;
+use Konekt\Address\Models\ZoneProxy;
+use Konekt\Address\Models\ZoneScope;
 use Konekt\Concord\BaseModuleServiceProvider;
 
 class ModuleServiceProvider extends BaseModuleServiceProvider
@@ -29,13 +37,28 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
         Country::class,
         Organization::class,
         Person::class,
-        Province::class
+        Province::class,
+        Zone::class,
+        ZoneMember::class,
     ];
 
     protected $enums = [
         AddressType::class,
         Gender::class,
         NameOrder::class,
-        ProvinceType::class
+        ProvinceType::class,
+        ZoneScope::class,
+        ZoneMemberType::class,
     ];
+
+    public function boot()
+    {
+        parent::boot();
+
+        Relation::morphMap([
+            'country' => CountryProxy::modelClass(),
+            'province' => ProvinceProxy::modelClass(),
+            'zone' => ZoneProxy::modelClass(),
+        ]);
+    }
 }